Refactor complete to change to active record methods.
