+++ app/views/navbar.erb
@@ -1,9 +1,6 @@
 <nav class="navbar navbar-inverse">
   <div class="container-fluid">
-    <ul class="nav navbar-nav">
-      <li><a href="/">Bike Share</a></li>
-    </ul>
-    <ul class="nav navbar-nav navbar-right">
+    <ul class="nav navbar-nav navbar-left">
       <li><a href="/stations">Stations</a></li>
       <li><a href="/trips">Trips</a></li>
       <li><a href="/conditions">Weather</a></li>
