--- capybara-201612052207073167351374.html
@@ -1,151 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <link rel="stylesheet" type="text/css" href="http://yui.yahooapis.com/3.18.1/build/cssreset/cssreset-min.css">
-  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
-  <link href="http://netdna.bootstrapcdn.com/twitter-bootstrap/2.2.2/css/bootstrap-combined.min.css" rel="stylesheet">
-    <link rel="stylesheet" type="text/css" media="screen" href="http://tarruda.github.com/bootstrap-datetimepicker/assets/css/bootstrap-datetimepicker.min.css">
-  <link rel="stylesheet" href="/application.css">
-  <title>Bike Share</title>
-</head>
-<body>
-  
-  <div class="container">
-    <div class="container">
-  <h2>Trips</h2>
-  <div class="list-group">
-    
-      <a href="/trips/1" class="list-group-item">
-        Start: 1930-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/2" class="list-group-item">
-        Start: 1931-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/3" class="list-group-item">
-        Start: 1932-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/4" class="list-group-item">
-        Start: 1933-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/5" class="list-group-item">
-        Start: 1934-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/6" class="list-group-item">
-        Start: 1935-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/7" class="list-group-item">
-        Start: 1936-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/8" class="list-group-item">
-        Start: 1937-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/9" class="list-group-item">
-        Start: 1938-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/10" class="list-group-item">
-        Start: 1939-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/11" class="list-group-item">
-        Start: 1940-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/12" class="list-group-item">
-        Start: 1941-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/13" class="list-group-item">
-        Start: 1942-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/14" class="list-group-item">
-        Start: 1943-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/15" class="list-group-item">
-        Start: 1944-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/16" class="list-group-item">
-        Start: 1945-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/17" class="list-group-item">
-        Start: 1946-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/18" class="list-group-item">
-        Start: 1947-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/19" class="list-group-item">
-        Start: 1948-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/20" class="list-group-item">
-        Start: 1949-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/21" class="list-group-item">
-        Start: 1950-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/22" class="list-group-item">
-        Start: 1951-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/23" class="list-group-item">
-        Start: 1952-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/24" class="list-group-item">
-        Start: 1953-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/25" class="list-group-item">
-        Start: 1954-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/26" class="list-group-item">
-        Start: 1955-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/27" class="list-group-item">
-        Start: 1956-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/28" class="list-group-item">
-        Start: 1957-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/29" class="list-group-item">
-        Start: 1958-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-      <a href="/trips/30" class="list-group-item">
-        Start: 1959-01-01 00:00:00 UTC, Station Not Found<br>
-      End: 2016-01-02 00:00:00 UTC, Station Not Found</br></a>
-    
-  </div>
-</div>
-
-<div class="container">
-  <ul class="pager">
-    <li class="previous"><a href="/trips/page/1">Previous</a></li>
-    <li class="next"><a href="/trips/page/1">Next</a></li>
-  </ul>
-</div>
-  </div>
-</body>
-  <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-  <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-</html>
